Image bugfix, bubble menu z-index bugfix, save notes using ctrl + s
